Basically a stash for trying to recover time when continueing. Doesn't work yet though

+++ src/timeHelper.h
@@ -23,6 +23,6 @@ StopWatch startTimer();
 
 void stopTimer(StopWatch *watch);
 
-double getTimeInSeconds(StopWatch *watch);
+int getTimeInSeconds(StopWatch *watch);
 
 #endif
